splitting up error reducer and chart reducer
