+++ stock-tracker/src/features/error/redux/actionTypes.js
@@ -1,0 +1,1 @@
+export const REQUEST_ERROR = "REQUEST_ERROR";
