+++ stock-tracker/src/features/error/redux/actions.js
@@ -1,0 +1,6 @@
+import { REQUEST_ERROR } from "./actionTypes"
+
+export const getErrorsAction = (requestName) => ({ 
+    type: REQUEST_ERROR,
+    payload: {requestName}
+  })
